Adjusted Tree.pm to call things in Page.pm.
Page.pm makes calls into the UseModWiki package to determine URLs
for links. It does not (yet) support all link types. It also
only allows control over the URL in a href, so the label of a
link cannot be controlled. This means that if a page does not yet
exist, the view on the web page does not show up differently
(indicating editing is needed for a wiki word).