Backed up the required version to 5.005.
This required change "our" to "use vars" etc.

Also the removal of "use warnings"

The split to a grep at line 329 in PurpleWiki::Parser::WikiText
throws a bunch of warnings about unitialized values. It would
be cool if we could figure out that little deal.

The tests pass, but this hasn't been functionally tested
(although the changes are essentially identical to the changes
I made on my running version).