A variety of cleanup changes. Latest functional tests continue to work, so committing, in case it is bed time.
The option to eval large code segments in long scalars has
been removed. The refactoring is in part to get rid of this,
and it is hard to work around while trying to clean up. So
it's gone. Moves some 'use' calls to the top.

Beginnings of moving global use vars variables to my as
possible.

Moved the instantiation of the wikiparser inside the
main loop slightly so it doesn't get called if some
error conditions occur. Bad me, this is early optmization with
very very minor gain.

Got rid of support for a cached index of page names. This
allows some functions to be a bit shorter.

Moved &InitLinkPatterns into the PurpleWiki::Config module
and removed as much as possible. The area is ripe for
further tuning as the patterns are not much needed since
they are replicated in the parser. They are used in the
UseMod package to confirm incoming URLs are good.

Removed support for page caching.

Removed functions that are not longer used because they
have been superceded by the parser. Including WikiLinesToHtml
and similar.

Took out the $FastGlob global variable and test so that
the FastGlob method we've been using all along is always
used.

Cleaned up Config.pm and config to reflect removed
variables.

Sigh. That was the easy part. The next stage is to deal
more effectively with all the global variables and the
functions that operate on those globals rather than
returning stuff.