Extensive changes to move all the code that does file manipulation into a separate module. This is a step in the right direction, from a figuring stuff out standpoint but is a long road to get where we need to go: I'm likely to trash just about everything I've done tonight now that I've identified the issues.
This checked in code is not functional for diffing
and revisions but works for displaying pages and
doing edits.

Basically what I did was copy a bunch of the OpenPage
and SavePage related functions to a module and then
do everything necessary to get rid of global variables.

The next step is now to create the classes identified
in that process. This is going to be necessary (or at
least makes the transition easier) to maintain the
versioning features that UseMod provides.

I have to say tonight's process would have been much
easier with a strongly typed language. A lot of my
errors would have been caught more directly.

I'll get to the rest of this around Saturday evening:
I didn't get the KB work done this evening because
my accounts with them weren't set up correctly, so
that's today's task.