Moved versioning into Database.pm (temporarily) and removed a bunch of code that's not required because the parser is present.
Got rid of UseModeWiki::CommonMarkup. This causes a slight
weirdness with diffing that we may wish to address:
CommonMarkup is the routine used to turn some text into
WikiHTML. I made it so that instead of sending output of a
diff to CommonMarkup it is sent through the parser routine
(with a separate parser instantiation). This allowed me to
get rid of a _lot_ of code. However if it's not already
there, I think we need a "don't purple this segment" option
for the Parser and Viewer.

Maintaining any sense of distance between the file handling
routines, the request handling routines, and the display
handling routines is quite a lump of coal. But I'm making
headway.

With this check-in diff stuff works, but seems to have some
trouble identifying major and minor revisions. I'm going to
let that bug sit while I do some other cleanups as I think
that fix itself eventually.

Removed a lot administrative commands including the
'maintain' action (points to &DoMaintain) and 'edit_links'
(points to &DoUpdateLinks) as as far as I know we haven't
used these things, and they are crufty in the way they tie
everything in knots. That cascaded into a bunch of removed
subroutines (also tied in knots).

Got rid of T() and Ts() where I found it.

Had to to dupe in ScriptLink and ScriptLinkDiff from wiki.pl
and then make a call back to the UseModWiki package for
ScriptName. That will be resolved later.

This commit mostly for safe keeping. Major revisions
forthcoming.