Further adventures in modularzing the PurpleWiki code. Making some headway. These changes are incomplete, I have to stop for fear of melting my brain. This message liable to be a mess.
Major change herein is that much of the code for processing
the Page, Section, Text and KeptRevision datastructures has
been moved into a Database directory and modules therein.
This further clarifies the data issues with all this stuff,
but what's there so far is once again not a complete deal.

Miscellaneous comments:

One again, killed off T() and Ts() routines. This time they
should all be gone, as I've removed the subs of those names
as well.

Page display is now handled via PurpleWiki::Database::Page
and friends.

Diffing sort of works, along with revision history display.

Saving is started but not complete.

These changes probably look rather complex, but I can tell
you that after working with them for a few hours, they are
much easier to get the brain around then the old way. I'm
especially pleased with the way I worked out to do
serialization of the datastructures for eventually saving to
disk.

You'll note that what I've done here is not very much like
what Eugene posted as a potential API in:

  http://purplewiki.blueoxen.net/forums/devel/2003-01/msg00005.html#nid04

This is because the framework surrounding data IO in UseMod
is outstandingly knotty. To unroll it all I've basically
been moving it into an OO form. From there we should be able
to make some sense of it (some is starting to come).

Next step:

- Continue to clean up wiki.pl by making calls into
  Database::*
- Then, start cleaning up, with the intent to remove,
  Database.pm.
- Then, back into wiki.pl, looking for opportunities to take
  noisy subs and clean them up, whittling wiki.pl down to
  some level of clarity

And somewhere in there I need to move user info out.

Things that are broken right now:

- If you show history on a page, it lists the latest
  revision with the wrong info, I think this is showing that
  saves were not happening correctly before. That is, that
  metadata is not travelling correctly.

  It also lists an extra, null, revision.
- There seems to be some confusion on how major and minor
  come into play with diffing
- When creating a new page it doesn't get the proper initial
  text.
- In fact, saving from such a page doesn't work at all (this
  is new)

Anywho, despite my utter inability to type at the moment,
I'm starting to feel in the groove with this stuff and
should have some good things after a while.