Adds support for User data in its own class through PurpleWiki::Database::User.
There is a lot of repetition between User, Page, and
KeptRevision that might benefit from a move to a SuperClass,
later.

Still very dirty all around. I have a change of heart about
accessor style when doing User.pm. Can't decide if it is
better or worse than the others. I've been on this long
enough that I'm starting to play to entertain myself and
that's probably a little bit undisciplined.

PurpleWiki::Database is now down to diff methods and general
file saving classes.

wiki.pl keeps request and cookie handling, link formatting,
user interaction, and functional pages such as RecentChanges
and User preferences.

There are things like &UserCanEdit and &ValidIdOrDie in
wiki.pl that are sort of user related, but not user database
related. I'm not sure what to do with them.

There are some time functions that can probably move either
to their own module or replaced outright with a third party
module.

&GetParam is a confusing method as it combines what is found
in the CGI request with what is in the User data to get a
result.

&GetRemoteHost is done in a old school way. Does it matter?

&DoOtherRequest: I've removed some of the functionality
represented here, but I'm not sure I've cleaned it all up.

&UpdateEmailList has a weird if conditional at the start
that I don't understand. I'm fairly sure it is just wrong,
anyone have some ideas on what's being tested there?

You'll not that I never did any unit tests for any of this,
that is because I had no idea what the functions are, either
conceptually or functionally. Once we have this version
functionally tested we should make a suite of unit tests for
it as that will definitely ease further refactoring.

My next step is to get revisions working. That's still a
problem. An additional symptom is that the first save of a
new file does not work, but the second does. So that's a
clue.

As I've said before: what's done so far is not meant to
pretty but to move things into chewable chunks. wiki.pl has
gone from 4055 lines to 2232. The new code out in modules
(seven of them so far) is 1640 lines.

Input on the questions above desired as well as these:

- Does it make sense to move diff stuff to its own module?
  It's not as receptive to being made into a clas
- What about request/cookie handling?
- Presentation is spread out all _over_ the place. Any
  suggestions on unifying that.
- Any general suggestions on chunking things?

(I'm looking forward to getting this stable so we can clean
it and then get on to things like unique id generation.)