No more global vars in wiki.pl. Anything that's needed throughout is now a my. Some of those can probably be adjusted to call data from the User reference (noted in the comments).
Moved the creation of the wikiparser back out to outside
edge of the package as it only ever needs to be created
once (thinking ahead to mod perl and the like).

Moved GetDiffHTML, ColorDiff, DiffToHTML back into wiki.pl
as they are presentation oriented. Removed them from
Database.pm

Cleaned out some prints to STDERR

As far as I can tell this is fully functioning. There's a
test installation at
http://www.burningchrome.com:8000/~cdent/wiki.cgi that has
some midnight ramblings in it.

Sigh, there are still 89 subs in wiki.pl. I'm checking now
for things that aren't used with the magic of shell scripts:

  for i in `grep ^sub wiki.pl|awk '{print $2}'` ;
  do j=`grep -c $i wiki.pl`; [ $j == "1" ] && echo $i &&
   grep $i PurpleWiki/*.pm PurpleWiki/*/*.pm;done

More to come.