Added some comments that will lead eventually to POD.
Fixed a bug in parsing the KeptRevision file. UseMod was
writing data as 'SEPDATASEPDATA'. I had been writing it as
'DATASEPDATA'. Turns out UseMod was just checking for empty
Sections whenever it created the list of stuff. I added a
similar thing and changed serialize() to use the old style.
Discovered this by importing the Helium wiki in as my new
test data.