This is on the nidGen branch.
Adds simple, bonehead support for a sequence generator for
NIDs. PurpleWiki::Sequence wraps a text file containging
the most recently used NID and has code to increase the
NID. This simpleness means no extra stuff required (Database
etc).

PurpleWiki::Parser::WikiText has been adjusted to not worry
about lastNid or biggestNid but instead just get the next
NID out of the sequence and use it.

A t/sequence is in place which makes sure the 9 to A and Z
to 10 carry overs work properly.

See it in action at:

  http://www.burningchrome.com:8000/~cdent/wiki.cgi?FrontPage

Issues:

- a six digit nid is presented with seven digits (the traditional
  extra zero on the front). good or bad?

- When working with existing data, [lastnid] now shows up
  because I didn't clean up that part of the parsing.

- The increment routines in Sequence seem rather heavy handed
  and are very implementation specific

- the $this versus $self stuff in Parser::WikiText confused
  me such that I couldn't figure out how to add a data member
  on to the object (I wanted $self->{sequence}). So I created
  it class level for now.

- There are few places in Parser::WikiText that were checking for
  valid NID form with \d. I replaced these with [A-Z0-9].

- There's a great deal of die()ing rather than graceful failures
  in Sequence

- The locking is simplistic

- It's slow

- An abandoned preview throws away the unused NIDs. I like it
  this way, but I think we may not have resolved that before.

- I don't know how to test for intentional failure (a die) when
  using Test.pm

- There exists a lot of cut and paste code in Parser::Wikitext

- I did this on the old Parser code, I'm not sure if that makes
  a difference or not.