This version works my wiki and rebuilds my blog, albeit with some unitialized values on concatenation, which I'll fix in a moment.
This commit corrects a bug in the wikihtml viewer where indents
were being closed as </indent> rather than </div>. They are
opened as <div class="indent">

It also changes the regular expression for matching WikiWords
with attached NIDs to match on the current spelling of NIDs.
As this is likely to change, we should probably pull that out
to a variable of some sort.

Which raises the spectre of config management which is a big
cow in all of this.