Fixes some problems with undefined and/or unitialized values
In Page.pm this is related to the $MainPage variable
associated with the MainPage/SubPage constructs. Apparently
there are times when the parser sees a / but has not MainPage
set. I've noted that this is funky in the code with a FIXME.

Elsewhere the issue is that we are building strings through a
tree of subroutine calls. Some of those calls look like this:

   return &foo . &bar

and one or the other of those returns doesn't return a string.

Over the long run we need to come up with some kind of idiom
for dealing with that sort of thing. I made an empty string ('')
available where necessary to clean things up but I'm not
sure this is ideal.