Okay, all that earlier merging and stuff was so I could justify this:
  http://www.burningchrome.com:8000/~cdent/wiki.cgi?FoodTest

If you look at edit text on that page, you'll see how it works.

This is prototype in the extreme. It is very slow because it is loading
web pages and parsing them to get at the nodes. So the FoodTest page
is loading three other pages _every_time_it_loads.

More in a separate message.

Details of changes:

- adds PurpleWiki::Transclusion which is the code for sucking
  text from web pages when provided with a nid. It is crufty, using
  some hardcoded data files. It is responsible for the whole string
  that is brought in, it is called from the transclusion main handler
  in the wikihtml eventhandler. all happens in main so state doesn't
  have to be maintained
- adjusts the Parser/WikiText to accept \[t[A-Z0-9]+\] as an inline
  node. we'll need to make this configurable
- make changes in the eventhandlers for text, wikitext, wikihtml
  and debug. only wikitext and wikihtml have been tested at this point

transclusion should probably be a configurable parameter like
wikiwords and freelinks.