purple.pl is the MovableType plugin for integrating PurpleWiki into the scene. It now requires no patching of the movable type distribution: everything can be controlled from this file or from the movable type config. There's some more info on it here:
  http://www.burningchrome.com:8000/~cdent/wiki.cgi?PurplePluginInstructions

(those are just notes, it doesn't actually work like that yet).

The lack of patching is done by overriding existing MT methods
by violating typeglobs. This was Ben Trott's suggestion, so
even though it is kinda nasty, it is blessed.

The major stumbling block now in place is config management: we
need to be able to create and pass around clean config objects
with different data directories as discussed here:

 http://purplewiki.blueoxen.net/forums/devel/2003-05/msg00070.html

Without that things are hard to do. Right now, for exaple I have
a bit of hard coded ugliness that provides the URL of the wiki
when it isn't the wiki itself calling things. Bad, that needs to
be in the config object somehow.