Moving in the direction of config cleanups.
Adjustments in wiki.pl to remove use of $NotFoundPg. If a link
leads to a nonexistent page, we want the user to be presented
with a blank page and the opportunity to edit, not a generic
"you said something weird page". Otherwise things like links
between chat logs and wikis don't work well. Some (Bill Seitz)
have complained about orphaned pages, but I convinced him
otherwise by suggesting emergence is good in this case.

Adjusted post editing HTTP redirection handling in GetRedirectPage
to remove some cruft that was in place for situations where CGI.pm is
not being used. We use CGI.pm elsewhere without checking this
so I figured may as well make CGI.pm rule here too.

When parse() is called on the purple wiki parser in wiki.pl, pass
the value of the $FreeLinks variable as the freelink parameter,
since it now exists.

Fixing an error in diffing Database.pm. I was passing strings
as straight strings rather than references to scalars, so it
was looking for and not finding files. Fixed.