Adds external modules requirements to Makefile.PL. I went a bit overboard as I added things that are likely to be in the perl distrubtion, but I figured better safe than sorry.
Also moved the use of Data::Dumper in Config.pm into the sub
that uses it because that sub is never called (only there for
debug purposes). There are a few other places where this might
be a wise thing to do.

Next: finding all global variables and killing them. But first
buying appliances.