cleanups to t/database.t to make sure it cleans up after itself correctly (actually just realized it needs more)
addition of t/page.t for testing the creation, saving and loading
of a page. this sort of recapitulates the page generation process.
it's not complete but hits the high points

Note: doing this makes me realize that although the file management
stuff have been moved out to its own set of classes, it is by no
means cleaned up. The UseMod method of doing things is still in place,
but we can change it more easily.