Adding a script to convert 0.1 installations to 0.2. What this script does is only deal with nid changes. It does not deal with config changes. In its comments it suggests a way to adjust the config (by hand).
This script gets a list of all the WikiPages, then for each page
gets the existing test, strips nids, strips lastnid, generates
a url for the page, passes the text back through the parser, gets
a new view with nids added, saves the text.

I tested this on a copy of the PurpleWiki wiki content and it worked,
after some tweaking.