This commit addresses the problem with infinite loops in transclusions
It gathers enough information to have both the source and dest url
at the point of transclusion. If they

 - the dest is the wiki url
 - and the source and dest are the same

return an error string.

It checks first for the wiki url because same source and dest on the
blog is _very_ handy

Please check over this, as I'm not sure it's quite right or the best
it could be.