Fixed issue 24: InterWiki not working with non-WikiWords and InterWiki subpages not working.  Turns out that the UseMod parser is much more lenient about InterWikiWords than the PurpleWiki parser was.  For InterWiki:PageName, PurpleWiki required that "PageName" be a WikiWord. UseMod does not.  UseMod allows InterWikiWords like InterWiki:pagename.
I reluctantly revised the PurpleWiki parser to also use this loose
InterWikiWord parsing.  I also caught a bug -- granular links to the new
style NIDs would not parse correctly.  I included tests for the former
bug, but not the latter (I just realized).  I'll do tests for the latter
bug and check those in next.

There seems to be a problem with subpage parsing in general.  v0.1 doesn't
have this problem, so it was introduced somewhere in-between.  Gonna
investigate that now as well.