Fixed a bug caused by moving some functions from wiki.pl into Page.pm. There was a global variable called $MainPage in wiki.pl that was used to store the name of the current page.  The problem was that when the functions were moved into Page.pm, the variable was not.  The result was that links to subpages (those darned subpages again!) weren't pointing to the right place.
Part of the reasons for moving the functions into Page.pm was to remove the
reliance of Page.pm on package UseMod.  So, to fix this bug, I created a
global $MainPage in Page.pm and had wiki.pl manipulate that variable.  Ugh.

Chris noted a few FIXMEs related to $MainPage.  I think this fix should
take care of those FIXMEs; Chris, please review these changes to confirm
this.

Also changed the default $CONFIG_DIR in wiki.pl to "/var/www/wikidb", an
imaginary but neutralish directory name.