This fixes a bug with parsing for transclusion. See:
  http://purplewiki.blueoxen.net/cgi-bin/wiki.pl?SandBox#nid6B

parser08 adds test cases for parsing for transclusion.

For reasons I don't quite understand, using a /x style regular
expression for the split of inline nodes was causing transclusions
that were inline with some other text (other than a : for indent)
to not be identified. After much experimentation I discovered that
collapsing the regular expression and getting rid of the /x
fixed it.

Sev.