Adjustment to Transclusion.pm to address a fun little bug:
Say you have a wiki page on which is created the nid '3'.

The you transclude it somewhere. Yay, it works.

Then you remove nid 3 from the original page.

The transcluding page, when displayed, will then suck up
the entire page and transclude it, not just the chunk desired.

The fix checks to see that the regular expressions actually do their
work and if they don't, provides a little warning message that
goes in the transcluded spot.

Also added the HTTP error code to the 'unable to retrieve content'
message as that would be handy to have at times.