Another PurpleWiki::Search module:
Search arts files through the wiki

Crufty because arts has namespace issues.

This implements a feature I'd love to have in the MovableType
version but because of the way data is stored, will require some
more complicated searching than current done.

The feature is that the NID of the text segment in which the query is
(first) found is appended to the URL that is stored. So you can jump
right to the place in the text.

I may go back and add this to the Wiki module.

As usual for my stuff: this is not perfect, definitely proof of
concept but I'm hoping worse is better than nothing in this case.

Irclogger in a bit.