Adding more detailed pod to the various search modules. It's still not complete, but certainly much better. I suppose somewhere there needs to be a enough information that explains how to create a new SearchModule. Is there enough there already?
Also added google stuff to etc/config as I had forgotten that before.

And, in another Rawlinsism, I added config() to Interface so that
queries to the config object in the search modules are by method
rather than field.

I realize this is also not perlish (for some) but it is the Agile
way to do it.