Updating VERSION in all the files (is there an automatic way to do this with MakeMaker or something?)
Adding information to UPGRADE, README and ChangeLog, each needs
some tuning to reflect the addition of WikiWhiteboard functionality.
I wasn't quite sure what to say.

Also, does there need to be a sort of "oh, hey, by the way, this
new stuff is cool" blurb anywhere? If so, where, or is that better
served by text associated with the announcements?