Adding Apache1Handler.pm (will get 2 in a bit). Adding it to the manifest as well.
Incorporating a few changes into Transclusion.pm:

* If for some reason no url is found in the index for a given nid
  request, return content in the transclusion saying so

* Change the || in the tie statement to 'or'. || and , fight for
  precedence, so when there is an error, it doesn't fly.

* Made the die on tie into a warn because we don't want things
  to die when transclusion fails, we simply want things to go
  slightly wrong (also I have a situation were it will fail, every
  time and I need to soften the blow until the cause is fixed)

Adjusting the call to the viewer so that errors cascade up the
stack. It seems that if you call the viewer in an eval, that
traps the errors (such as those that might be happening in a
transclusion) without reporting them. We need to be able to see
those errors, especially while we are in development, so I
switched things around a bit. I've now realized I could have
rethrown @$ but the way I've done it accomplishes the same
thing. There may be better ways, please look.

Apache1Handler needs documentation.