Adding basic documentation to the Apache handlers.
Adding info on them to the changelog.

It's getting to the point where purplewiki needs a howto or something
along those lines. Something that describes how to use it as more
than just a wiki. I'm suppose I'm probably the man for that job.
If anybody has any suggestions on how to start and what topics
to cover, please let me know.