THIS COMMIT IS ON A BRANCH called remoteSequence
The code in this commit enables remote nid handling. The primary
feature of which is remote transclusion.

 http://www.burningchrome.com:8000/~cdent/wiki.cgi?SandBox#nid2KA

The content at the next nid is remotely transcluded.

(you won't be able to follow the link because the other end is inside
an authenticated site that my wiki machine has access to)

THIS IS PROOF OF CONCEPT CODE, I DON'T EXPECT THE INTERFACE
TO BE STABLE NOR DO I YET KNOW IF IT IS ANY GOOD.

It's certainly nice and slow.

Config.pm adds the RemoteSequence value which points to a
URL where an apache handler (Apache1NidHandler (not yet tested) or
Apache2NidHandler (running)) performs the following two functions:

* given a <count>/URL pair, return count NIDs (the count functionality
  is in the backend, but not implemented in the frontend)

* given a nid, return the URL it is found on

Parser/WikiText.pm passes the RemoteSequence value to Sequence::getNext()
If it is null, a local sequence is used, otherwise, LWP is used to
to send the provided URL off to the remote sequence and get back the NID.

Sequence presents its usual interface but branches if RemoteSequence
has been passed to it. It also has a new method getURL.

Transclusion.pm uses getURL to determine the URL it should look at
for pulling in the content at a particular NID.

Please make some arrangements me with me if you are willing and able to
create some purple numbers from a shared sequence with me. I think
there are going to be plenty of issues to shake out.

This code was remarkably easy to fit in. In think Rawlins gave me an
education of something.