This is a version of mhpurple.pl that supports a sequence or remote sequene. It ads the REMOTESEQUENCE and SEQUENCE custom variables (see the README). I did this rather than a wiki config dir as it allows for a sequence that is shared amongst wikis.
Both LINKBASE and BASEURL are required as the linkbase sub
has the code for generating the messageURL which is require
for passing to the sequence.

I've tested this with both the SEQUENCE and REMOTESEQUENCE.

Eugene, I'll leave it to you to do what is required to make
it comply with newer mhonarc.

I've decided to include it in the extras dir as I that seems
the most reasonable place for the time being.