This is incomplete, but works.
This adds support for transclusions within the wiki by using the 
view drivers rather than a screen scrape.

wikihtml.pm requests a transclusion normally when it runs across one 
in its tree.

Transclusion.pm does some guessing to determine which style to
try. If it thinks it can go internal it loads the page, makes a 
tree, and views it using subtree to return the subtree to wikihtml.pm

wikihtml.pm checks to see if what's returned from Transclusion is
a ref or a string. If it is a ref it incorporates it into the
tree it is parsing and traverses that section of the tree before
continuing where it left.

if it is a string, it just adds it to the output string.

I added a class field to InlineNode so that I can declare the 
class of html items created.

For the future I think it might be valuable to consider an object
hierarchy of Transcluders, somewhat along the lines of the
Search::InterfaceS (and probably using much of the same config
info). This will allow us to retrieve transcludes from irc logs, 
arts files, blog entries, etc, without needing to go to the network
if the transclusion is local.

This has two implications:

# We need a good way to know if content is local. I'm not sure if
  we always have access to the servername without setting it by
  hand in config.

# We need a remote transclusion service, much like the remote nid
  service that will request transclusions from remote services
  transparently (much like the way RemoteSequence works now), yet
  still fall back to screen scraping when desired.