Adjusts Sequence generation so there can be one sequence source for multiple wiki configurations on the same machine by adding a LocalSequenceDir config value that defaults to the DataDir
In the process renamed RemoteSequence to RemoteSequenceURL so it
more parallels the local sequence name.

What this allows, or the reason I did this, is so that places 
like BlueOxen, that have a big pile of wikis, can, if desired,
transclude amongst themsevles while still maintaining some 
separation. 

This complicates the transclusion scenario somewhat. Doesn't break
anything (yet, that I can think of) but makes "going local"
harder to determine.