Created a rawtext driver which doesn't do any text wrapping or special formating like that.  Doesn't include header or footer information either.
This is useful to getting raw access to a text based form of the node for 
things like transcluding into IRC via a purple number aware bot.

rawtext is the parent of the text driver.

I tested both text and rawtex a few times, they seem to work.  In reality
there isn't much of a functional change.