I needed/wanted this for some stuff I was doing, so now it exists.
PurpleWiki::Rss creates an RSS 1.0 string of the 15 most recent 
changes (can do more if desired).

rssCreate uses it to create a full content feed of the recent changes

Turns out this is delightful to have in a good rss aggregator like
NetNewsWire. Especially since the PurpleNumbers are clickable and 
proper.

There's an issue that we need to resolve though: we don't have a
good definite way for retrieving the full url of a page. 
Sometimes ScriptName is the FullURL of the wiki, sometimes FullURL
is. 

I used ScriptName in this stuff as that's what's correct in my 
setups. But I note that Matthew tends to go the other way in 
some of his stuff. We need, presumably, a single method that will
do the right thing whatever is set.