Created a Config Singleton branch.  Nothing uses the Singleton magic, but I wrote a mod_perl test (which I didn't include) that seems to show it holds up.
This stuff depends on Apache::Singleton, which means it introduces an external
module dependancy.  Apache::Singleton works with or without mod_perl, so it
doesn't cost us anything to use it.  It's also so brain dead simple it could
be reimplemented with very minimal effort, so we want to avoid the dependancy
then that's always a good option.