Corrected a bug with overzealous hardrules.  As reported by Chris Dent:
  http://purplewiki.blueoxen.net/forums/devel/2004-03/msg00043.html

We need automated tests for view drivers.  The funkiness with the hard
rules makes this especially important.  It's not enough for the parser
to parse WikiText hardrules correctly, because it does something
nontrivial with them.  Thus, converting the data structure back to
WikiText or anything else for that matter is also nontrivial.  Thus,
we need tests for both the parsing (which we have) and the output
(which we don't).