Added multiline support for list items and indented paragraphs.  See the thread:
  http://purplewiki.blueoxen.net/forums/devel/2004-02/msg00124.html#nid06

for background.  Basically, this makes the parser a bit more strict
about newlines being terminators for nodes.  In other words, if you have:

  * list item.
  should be part of above list item.

all of the above is part of the same list item.  Before, the second
line would have started a new paragraph.  The continued line may also
be safely indented, if that's your cup of tea.

One pleasant side-effect of all this is that the code actually became
slightly shorter/cleaner, because it removed a special case that
probably wasn't being used anyway.