Modified the parser to support the XHTML model for nested lists.  See:
  http://purplewiki.blueoxen.net/forums/devel/2004-02/msg00113.html#nid04

for details.

An unimportant observation: We now have unprecedented behavior for a
StructuralNode -- node types (li, dd) that have both content and
children.  The data structure was flexible enough to support this, so
it wasn't a problem.  (This doesn't take into account the recent
discussion regarding eliminating the distinction between Structural and
Inline nodes.)

I moved the list tests out of tree_test01.txt (and parser01.t) into
tree_test14.txt (parser10.t).  It makes the tests a bit more modular,
and it was easier for me to play.  The tests have been updated to 
support the new parsing scheme.

The trickiest part was modifying the View drivers.  There was an
assumption in the debug driver regarding indentation that is no
longer valid.  Indentation now occurs every time a child is seen, which
is probably how we should have done it in the first place.  Both the
wikitext and wikihtml drivers needed to be modified to generate NIDs in
in the *Main methods, not in the *Post methods.  Otherwise, the NIDs go
in the wrong place (after the sublist rather than after the list item
content) and you get funky newline behavior in the wikitext output.