Fixes a typo in header processing.
This address the problem were subsequent headers were getting
smaller and smaller despite the number of = chars used.

HOWEVER: this aggrevates a problem with PrevSection* that I'm unable
to address with a cursory look: spurious <hr> (and ---- in wikitext)
are being introduced. See:

  http://www.burningchrome.com:8000/~cdent/wiki.cgi?SandBox

I didn't put in the <hr>s between the Header # lines.

* PrevSection's meaning/purpose/destiny is not immediately obvious.