First commit of templatization changes.  There are a new set of PurpleWiki classes: PurpleWiki::Template::*.  The default template driver is PurpleWiki::Template::TT.pm, which uses Template Toolkit (http://www.tt2.org/).  New template drivers can be created by inheriting from PurpleWiki::Template::Base.pm.  It should be very straightforward to use other templating systems, although I haven't actually tried it yet.
Templates are stored in wikidb/templates.  I haven't completed the
templatization process yet, so there are other templates that still
need to be developed.  I use the file extension .tt, which is hardcoded
in PurpleWiki::Template::TT.  This allows you to use only the basename
of the template file in the template call, which allows you to easily
switch templating systems by using a different PurpleWiki::Template
class.

I moved etc/intermap and etc/config into wikidb.  This makes more
sense intuitively for installation; administrators can simply copy the
wikidb directory over to wherever they want it (or we can modify
Makefile.PL to do it for them).  Updated the installation directions
in README accordingly.

The only other class that was touched was PurpleWiki::Database.pm.
I added a recentChanges function, a modified version of Chris's 
PurpleWiki::Rss::_getRecentChanges method.  At some point, we should
modify that class to use the new recentChanges function.  The new
function fixed a few minor bugs in Chris's code and generates a data
structure that is a bit more convenient for the viewRecentChanges
template.

Modified wiki.pl substantially, although many more modifications await.
In the process of templatizing various functions, I removed functionality
that didn't make sense to me.  Some of this was in the UI.  For example,
in UseModWiki (and prior versions of PurpleWiki), when you clicked on
RecentChanges, it gave you a link to display recent changes starting
from the current time.  That's totally unnecessary; you can just reload
the page.  There are subtler, more egregious examples of this throughout.

I didn't change the UI substantially.  It looks mostly like the old UI
for the most part.  I added a "Pages Index" to the default header, and
removed the nav bar from the footer.  It should be XHTML compliant except
for the headers.

After I templatize the remaining functions, I'm going to carefully
prune out the remaining wiki.pl functions and config variables that are
no longer necessary.  I'll also make the templates XHTML compliant, 
examine the variables for user-friendliness, and modularize the templates
where appropriate (for example, the nav bar should be in its own file).