Changing how version numbers are handled.
The global distribution number if kept in Makefile.PL, it has
been changed to 0.93 for the time being.

$VERSION in all the files have been changed to the subversion
revision in which they were last changed.

I'm not sure of the exact impact this will have on CPAN, but it
should be okay.

This means we don't have to manage $VERSION in each file now but
still have meaningful information in there.