changes to get wiki.pl operational on hot
A brand new wikidb has no rclog file, so the rss function in
Database.pm (and the one in Rss.pm it was copied from, so presumably 
the one it was copied from in the WikiRPC lib) dies. I've
changed it so that if the rclog can't be opened, there's just
no rc stuff. 

wiki.pl was using the ScriptLink config value rather 
than ScriptName