This commit merges Eugene's template changes to the Singleton config so I can safely test them. Because of all my id changes recently this was a very painful merge. I don't recommend mergine the trunk in the future to a branch in the past. It makes a big mess.
Anyway, the singleton stuff works properly with alongside the
templates. It seems like a fine implementation. It doesn't 
currently support multiple wikis on the same installation and
binary, but this change will make that easier, I think.

It's getting late enough now where I'm becoming fog.

Eugene if you don't think this is going to mess up your 
template, could you go ahead and merge it to the trunk? I
don't want to do it myself right now as I've not got
the proper wakefulness to decipher svn's magic.