Here's a working but not fully functional HTML::Template implementation for the template system.
It's not really all that pretty: I've just replicated what Eugene had,
moved some things into subtemplates, and then set about noting 
where things fell down. Those notes are in the form of FIXME coments
in the *.ht template files and one in wiki.pl.

Another difference is that I've started adding classes or ids to things
as an experiment, and I'v removed <hr /> where I've found it, preferring
boundaries on divs and things like that.

I have a design in progress on

  http://www.burningchrome.com:8000/~cdent/tw.cgi

to further experiment with things. I'm really only working with the 
viewPage template at this point as far as fiddling with things.

I figured, at least for the time being, that I would put the 
template files in the same directory as is currently used and
give them an ht extension.

I don't consider this done or anything like that, in fact I suspect 
its primary worth will be as a tool to track the functionality in the
TT templates and see what is or is not "okay".

I'll follow up this with some comments on the fixmes.